requirements.txt
fastapi>=0.128.0
langchain-community>=0.4.1
langchain-openai>=1.1.7
langchain-pinecone>=0.2.13
langgraph>=1.0.6
pinecone>=7.3.0
pypdf>=6.6.0
uvicorn>=0.40.0
pydantic-settings>=2.0.0
python-multipart>=0.0.9
